requirements.txt
fastapi
uvicorn
stripe
psycopg2-binary